added fdin for simple redirection <
